+++ lib/src/funcs/funcs.dart
@@ -1,3 +1,4 @@
 export 'format_funcs.dart';
+export 'logs.dart';
 export 'parse_funcs.dart';
 export 'tool_funcs.dart';