+++ CHANGELOG.md
@@ -11,4 +11,13 @@
 # 0.1.0+4
 
 - Add formatFullDateTime func
-- Fix typo of name func.
+- Fix typo of name func
+# 0.1.0+5
+
+- Add compressFile func
+- Add static for some funcs
+- Add Logs class
+- Add log error for cache funcs
+# 0.1.0+6
+
+- Add stringToDecimal6 func